% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\docType{package}
\name{mycensus}
\alias{mycensus}
\alias{mycensus-package}
\title{mycensus: A package for easy compilation of census estimates for the Puget Sound Region.}
\description{
This package provides functions to easily collect census estimates across geographies within the
Central Puget Sound region.
}
\author{
Christy Lam \email{clam@psrc.org}
}
