% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct020af-data.R
\docType{data}
\name{reg_dec_pct020af}
\alias{reg_dec_pct020af}
\title{Decennial 2010 PCT020A-F Group Quarters Population by Group Quarters type for the Central Puget Sound Region}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 960 rows and 7 columns.
}
\source{
Packages: psrccensus via tidycensus
}
\usage{
data(reg_dec_pct020af)
}
\description{
Decennial 2010 PCT020A-F Group Quarters Population by Group Quarters type for the Central Puget Sound Region
}
\keyword{datasets}
